+++ GenericServices/Internal/Decoders/DecodedDto.cs
@@ -39,8 +39,8 @@ namespace GenericServices.Internal.Decoders
             _matcher = new MethodCtorMatcher(publicConfig.NameMatcher);
             _perDtoConfig = perDtoConfig; //can be null
             LinkedEntityInfo = entityInfo;
-            if (entityInfo.EntityStyle == EntityStyles.HasNoKey)
-                //If HasNoKey then exit immediately as properties etc 
+            if (entityInfo.EntityStyle == EntityStyles.HasNoKey || entityInfo.EntityStyle == EntityStyles.OwnedType)
+                //If HasNoKey or is OwnedType then exit immediately as properties etc 
                 return;
 
             PropertyInfos = dtoType.GetProperties()